install virt-what and use facter's default virtual facts
Discussion in #d-a suggests that on jessie facter needs the virt-what
package installed for the virtual and is_virtual facts to work.
Install it on jessie (and purge on stretch and later), and revert to
using the facter's default facts about virtualization.

Also remove our own systemd based virt fact.